+++ CHANGELOG.md
@@ -68,6 +68,9 @@
 - Changes to SqlServerMemory
   - BREAKING CHANGE: Parameters SQLServer and SQLInstanceName has been renamed
     to ServerName and InstanceName respectivly ([issue #308](https://github.com/PowerShell/xSQLServer/issues/)).
+- Changes to SqlServerNetwork
+  - BREAKING CHANGE: Parameters SQLServer has been renamed to ServerName
+    ([issue #308](https://github.com/PowerShell/xSQLServer/issues/)).
 - Changes to SqlServerOwner
   - BREAKING CHANGE: Parameters SQLServer and SQLInstanceName has been renamed
     to ServerName and InstanceName respectivly ([issue #308](https://github.com/PowerShell/xSQLServer/issues/)).